backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.7.4,<3.0.0
langchain-openai==0.2.10
langchain-core>=0.3.21
langchain==0.3.9
python-dotenv==1.0.0
httpx==0.26.0